# dc-api/openapi.yaml
openapi: 3.0.3
info:
  title: DC-API
  version: 0.1.0
  description: |
    WSO2 Infrastructure Platform Control Plane API.

    DC-API sits in front of a Rancher + Harvester infrastructure layer and
    exposes a clean, tenant-isolated REST API for managing VMs, RKE2 clusters,
    virtual networks, DNS zones, and RBAC — in the style of AWS / Azure.

    **Authentication**

    Every `/v1/*` endpoint requires a `Bearer` token in the `Authorization`
    header. Two token types are accepted by the same scheme:

    1. **Asgardeo OIDC JWT** — obtained via the `dcctl login` PKCE flow. The JWT
       must carry a `groups` claim containing at least one group matching the
       tenant group prefix (default `dc-tenant-<tenant-id>`).

    2. **Service-account token** — a long-lived opaque token with the format
       `dcapi_sa_<lookup_id>_<secret>`, issued by
       `POST /v1/tenants/{tenant_id}/service-accounts`. Store it on creation;
       it cannot be retrieved again.

    **Async provisioning**

    Long-running operations (VM create, VNet create, etc.) return `202 Accepted`
    immediately. The response body contains the resource object with
    `"status": "PENDING"` and a `note` field with the polling URL. Poll the
    GET endpoint until `status` transitions to `ACTIVE` or `FAILED`.

    **Status lifecycle**

    `PENDING → ACTIVE` or `PENDING → FAILED`

    Deletion: `ACTIVE → DELETING → (row removed on success)`

    **Error format**

...
servers:
...
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /v1/tenants/{tenant_id}/projects/{project_id}/vnets/{vnet_id}:
    parameters:
      - $ref: "#/components/parameters/TenantID"
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/VNetID"

    get:
      operationId: getVNet
      tags: [networking-vnets]
      summary: Get a VNet by UUID
      responses:
        "200":
          description: VNet resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VNet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteVNet
      tags: [networking-vnets]
      summary: Delete a VNet (async)
      description: |
        Marks the VNet as DELETING and triggers async removal from KubeOVN.
        Returns 409 if the VNet has non-FAILED subnets, route tables, or
        peerings — delete those first.

        **Required role**: `owner`.
      responses:
        "202":
          description: Deletion accepted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: VNet has active dependents (subnets, route tables, or peerings)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "VNet has active subnets; delete them first: 3a1c..."
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Subnets ───────────────────────────────────────────────────────────────────
...
  /v1/tenants/{tenant_id}/projects/{project_id}/vnets/{vnet_id}/subnets/{subnet_id}:
    parameters:
      - $ref: "#/components/parameters/TenantID"
      - $ref: "#/components/parameters/ProjectID"
      - $ref: "#/components/parameters/VNetID"
      - name: subnet_id
        in: path
        required: true
        description: UUID of the subnet
        schema:
          type: string
          format: uuid
          example: 3a1c2e5d-0000-0000-0000-000000000021

    get:
      operationId: getSubnet
      tags: [networking-vnets]
      summary: Get a subnet by UUID
      responses:
        "200":
          description: Subnet resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subnet"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteSubnet
      tags: [networking-vnets]
      summary: Delete a subnet (async)
      description: |
        Marks the subnet DELETING and triggers async KubeOVN removal.

        Returns 409 if any VMs, bastions, or clusters are still attached to
        the subnet (via `subnet_id` on the resource row). KubeOVN refuses to
        release the logical switch while LSPs are pinned, so attempting the
        delete asynchronously would leave the subnet stuck in FAILED. Delete
        the attached resources first, then retry.

        **Required role**: `owner`.
      responses:
        "202":
          description: Deletion accepted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Subnet has attached resources (VMs, bastions, or clusters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: "subnet has attached resources; delete them first: chalitha-bastion (bastion), web-01 (vm)"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Route Tables ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Accepts either:
        1. An Asgardeo OIDC JWT (RS256) — obtained via the PKCE flow (`dcctl login`).
        2. A DC-API service-account token with the format
           `dcapi_sa_<lookup_id>_<secret>` — issued by
           `POST /v1/tenants/{tenant_id}/service-accounts`.
  parameters:
...
    TenantID:
      name: tenant_id
      in: path
      required: true
      description: Tenant identifier (slug, e.g. "iam-platform")
      schema:
        type: string
        pattern: '^[a-z][a-z0-9-]{0,30}[a-z0-9]$'
        minLength: 2
        maxLength: 32
        example: iam-platform
    ProjectID:
      name: project_id
      in: path
      required: true
      description: Project identifier (slug, e.g. "platform-infra")
      schema:
        type: string
        pattern: '^[a-z][a-z0-9-]{0,18}[a-z0-9]$'
        minLength: 2
        maxLength: 20
        example: platform-infra
    VNetID:
      name: vnet_id
      in: path
      required: true
      description: UUID of the virtual network
      schema:
        type: string
        format: uuid
        example: 3a1c2e5d-0000-0000-0000-000000000010
  responses:
    BadRequest:
      description: Bad request — validation error or malformed JSON
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "name is required"
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "missing Authorization header"
    Forbidden:
      description: Insufficient role for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "role 'member' is insufficient; 'owner' required"
...
    NotFound:
      description: Resource not found (also returned for cross-tenant access attempts)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "resource not found"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "failed to register resource"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable description of what went wrong
          example: "name is required"
    ResourceStatus:
      type: string
      enum: [PENDING, ACTIVE, FAILED, DELETING, DELETED]
      description: |
        Lifecycle status of any DC-API resource.
        - `PENDING` — accepted, provisioning in progress
        - `ACTIVE` — running and healthy
        - `FAILED` — provisioning or deletion failed
        - `DELETING` — deletion requested, async removal in progress
        - `DELETED` — terminal; appears only as an audit event's to_status
          (resource records are removed at deletion, never parked here)

    # ── Projects ───────────────────────────────────────────────────────────────
...
    VNet:
      type: object
      required: [id, tenant_id, name, region, address_space, status, provider_type, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: 3a1c2e5d-0000-0000-0000-000000000010
        tenant_id:
          type: string
          example: iam-platform
        name:
          type: string
          example: prod-vnet
        region:
          type: string
          example: lk-dev
        address_space:
          type: array
          items:
            type: string
          example:
            - 10.10.0.0/16
        description:
          type: string
          example: Production VNet
        status:
          $ref: "#/components/schemas/ResourceStatus"
        provider_type:
          type: string
          example: kubeovn
        message:
          type: string
          description: Human-readable status detail or error message
        created_at:
          type: string
          format: date-time
          example: "2026-05-09T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-05-09T10:01:00Z"
...
    Subnet:
      type: object
      required: [id, vnet_id, tenant_id, name, cidr, status, provider_type, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
          example: 3a1c2e5d-0000-0000-0000-000000000021
        vnet_id:
          type: string
          format: uuid
          example: 3a1c2e5d-0000-0000-0000-000000000010
        tenant_id:
          type: string
          example: iam-platform
        name:
          type: string
          example: app-subnet
        cidr:
          type: string
          example: 10.10.1.0/24
        gateway:
          type: string
          example: 10.10.1.1
        description:
          type: string
        status:
          $ref: "#/components/schemas/ResourceStatus"
        provider_type:
          type: string
          example: kubeovn
        message:
          type: string
        created_at:
          type: string
          format: date-time
          example: "2026-05-09T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-05-09T10:01:00Z"
...